+++ .gitignore
@@ -67,3 +67,7 @@ forge-artifacts
 # Compiler files
 cache
 out
+
+# Presigning and deployment directories
+presigned-createx-deployment-transaction
+broadcasted-createx-deployment-transaction
