+++ scripts/presign.ts
@@ -10,7 +10,7 @@ const RED = "\x1b[31m";
 
 const dir = path.join(__dirname, "presigned-createx-deployment-transaction");
 
-export async function sign() {
+export async function presign() {
   try {
     if (!Array.isArray(accounts)) {
       throw new Error("No private key configured.");
@@ -37,15 +37,17 @@ export async function sign() {
     const CreateX = await hre.ethers.getContractFactory("CreateX");
     const initCode = await CreateX.getDeployTransaction();
 
-    // Prepare the replayable transaction payload
+    ////////////////////////////////////////////////
+    // Prepare the replayable transaction payload //
+    ////////////////////////////////////////////////
     const tx = new hre.ethers.Transaction();
-    tx.to = null;
-    tx.gasLimit = 3_000_000;
-    tx.gasPrice = hre.ethers.parseUnits("100", "gwei");
-    tx.data = initCode.data;
-    tx.chainId = 0; // Disable EIP-155 functionality
-    tx.nonce = 0;
-    tx.type = 0;
+    tx.to = null; // A contract creation transaction has a `to` address of `null`
+    tx.gasLimit = 3_000_000; // A normal deployment currently costs 2,543,595 gas. We can later add different `gasLimit` levels for the final presigned transaction
+    tx.gasPrice = hre.ethers.parseUnits("100", "gwei"); // A gas price of 100 gwei
+    tx.data = initCode.data; // Contract creation bytecode
+    tx.chainId = 0; // Disable EIP-155 functionality (https://github.com/ethers-io/ethers.js/blob/bbcfb5f6b88800b8ef068e4a2923675503320e33/src.ts/transaction/transaction.ts#L168)
+    tx.nonce = 0; // It must be the first transaction of the deployer account
+    tx.type = 0; // Set to legacy transaction type 0
 
     // Sign the transaction
     const signedTx = hre.ethers.Transaction.from(
@@ -95,7 +97,7 @@ export async function sign() {
   }
 }
 
-sign().catch((error) => {
+presign().catch((error) => {
   console.error(error);
   process.exitCode = 1;
 });
